Added directory for Intro components
